Combine 3 objc/swift unit tests into one swift class
